+++ packages/dumb/src/components/stats/average-daily-dose/average-daily-dose-stat.tsx
@@ -76,7 +76,7 @@ const AverageDailyDoseStat: FunctionComponent<AverageDailyDoseStatProps> = (prop
           </Box>
           <SimpleValue
             className={styles.insulinTitle}
-            suffix={t('U')}
+            suffix='U'
             value={formatDecimalNumber(dailyDose, 1)}
           />
         </div>