+++ packages/medical-domain/test/mock/data.statistics.mock.ts
@@ -26,12 +26,13 @@
  */
 
 import type DateFilter from '../../src/domains/models/time/date-filter.model'
-import { createMealData, createWizardData } from '../data-generator'
-import { type Meal, type Wizard } from '../../src'
+import { createMealData, createRandomBasal, createRandomBolus, createWizardData } from '../data-generator'
+import { type Basal, type Bolus, type Meal, type Wizard } from '../../src'
 
 const abbottDevice = 'AbbottFreeStyleLibre-XXX-XXXX'
 const dexcomDevice = 'Dexcom-XXX-XXXX'
 
+export const MS_IN_HOUR = 864e5 / 24
 export const bgDataSourceOneDay: Array<[Date, string]> = [
   // data for one day and two days tests
   [new Date('2018-02-01T00:00:00.000Z'), abbottDevice],
@@ -40,7 +41,18 @@ export const bgDataSourceOneDay: Array<[Date, string]> = [
   [new Date('2018-02-01T00:45:00.000Z'), dexcomDevice],
   [new Date('2018-02-01T00:50:00.000Z'), dexcomDevice]
 ]
-
+export const basalsData: Array<[Date, number, number, string]> = [
+  [new Date('2018-02-01T01:00:00Z'), 0.25, MS_IN_HOUR, 'automated'],
+  [new Date('2018-02-01T02:00:00Z'), 0.75, MS_IN_HOUR, 'scheduled'],
+  [new Date('2018-02-01T03:00:00Z'), 0.5, MS_IN_HOUR, 'scheduled'],
+  [new Date('2018-02-03T03:00:00Z'), 0.5, MS_IN_HOUR, 'scheduled']
+]
+export const bolusData: Array<[Date, number]> = [
+  [new Date('2018-02-01T01:00:00Z'), 4],
+  [new Date('2018-02-01T02:00:00Z'), 5],
+  [new Date('2018-02-01T03:00:00Z'), 6],
+  [new Date('2018-02-03T03:00:00Z'), 4]
+]
 export const bgDataSourceTwoDays: Array<[Date, string]> = [
   // data for two days tests
   [new Date('2018-02-01T00:50:00.000Z'), dexcomDevice],
@@ -82,6 +94,24 @@ export const buildWizardData = (data: Array<[Date, string]>): Wizard[] => (
     {
       ...createWizardData(wizardData[0]),
       deviceName: wizardData[1]
+    }
+  ))
+)
+export const buildBasalsData = (basalsData: Array<[Date, number, number, string]>): Basal[] => (
+  basalsData.map((basals) => (
+    {
+      ...createRandomBasal(basals[0], basals[2]),
+      rate: basals[1],
+      duration: basals[2],
+      deliveryType: basals[3]
+    }
+  ))
+)
+export const buildBolusData = (bolusData: Array<[Date, number]>): Bolus[] => (
+  bolusData.map((bolus) => (
+    {
+      ...createRandomBolus(bolus[0]),
+      normal: bolus[1]
     }
   ))
 )