+++ test/locales/translations.test.js
@@ -202,7 +202,17 @@ const ignoredTransKeyForYourLoops = [
   'no-medical-files-patient',
   'added',
   'deleted',
-  'updated'
+  'updated',
+  'no-medical-files-patient',
+  'average-daily-insulin',
+  'average-daily-insulin-tooltip',
+  'daily-dose-per-weight',
+  'total-insulin',
+  'total-insulin-days-tooltip',
+  'total-insulin-how-calculate-tooltip',
+  'average-daily-total-insulin',
+  'ratio-dose',
+  'weight'
 ]
 const ignoredTransKeyInYourLoopsFiles = [
   'yourloops|${s}',
@@ -340,7 +350,6 @@ describe('Locales tests', () => {
         missingTranslations.push(key)
       }
     }
-
     expect(unusedTranslations, 'Unused translations').to.be.empty
     expect(missingTranslations, 'Missing translations').to.be.empty
   })